Modify the way config handles are cleaned up when no longer needed. This fixed a problem where xauth and config acks were being confused for one another.
Correctly set the mask that defines the attributes to be sent in push mode. This simply gets set the mask to include all attributes configured for the source.

Correctly read the returned attribute list in a config push ack so we can log what was accepted by the peer. Remove an unused pointer in the iked base class.

git-svn-id: svn://svn.shrew.net/ike@29 1bb156eb-19af-db11-b919-0014c23ffdf2