Add support to iked for negotiating replacement IKSAMP SAs. This means the that the tunnel connection can now exist beyond the lifetime negotiated during phase1. A new soft lifetime expire event was added to initiate a replacement ISAKMP SA when the previous SA is close to expiration time. For now, the tunnel Xauth and config state are cleared to support gateways that require a full Xauth during the ISAKMP rekey process. NAT-T negotiation state was moved into the tunnel handle. This is to preserve state across ISAKMP SA negotiations.
Fix a bug in iked that was causing a crash. This occurred when a peer attempted to initiate a new ISAKMP SA negotiation with iked running in client mode. Part of the problem resolution involved moving the policy and address release code out of the phase1 cleanup handler and into the tunnel cleanup handler. This also gets us closer to supporting ISAKMP rekey attempts as a responder when iked is configured to act as client gateway. Bug reported by David Santinoli.

Modify iked to gracefully handle the condition where a route to the distant host is not available at connection time. Bug reported some time ago by Alex Funk.

Add a new client feedback message to iked that more accurately notifies the client when phase1 negotiations have failed. Previously the client reported that the peer gateway was not responding.


git-svn-id: svn://svn.shrew.net/ike@375 1bb156eb-19af-db11-b919-0014c23ffdf2