Modify iked to split the pfkey_recv_acquire message handler into two functions. The first function, still named pfkey_recv_acquire, just handles the message parsing and calls the second function named pfkey_init_phase2. The second function handles the phase2 setup and exchange initiation. This allows other code paths to initiate new phase2 exchanges by specifying the outbound policy id and policy type.
Add support for nailed connections to iked. When enabled for a peer configuration, iked will mark any related tunnel generated policies as nailed. When a policy add response message is received from SPD, negotiations will be immediately initiated for a new SA that matches the policy. The policy ID is cached in the phase2 handle so that when a phase2 soft expire event fires, a replacement SA is then negotiated. Any acquire messages received for a nailed policy are ignored as iked will manage exchange initiation internally.

git-svn-id: svn://svn.shrew.net/ike@388 1bb156eb-19af-db11-b919-0014c23ffdf2