Improve libith. Modify the event timer class to not use a polling event loop. It now sleeps until the next event should be executed or a wakeup request is received. A new class has also been added that provides provides a timed wait condition that can be woken by another thread.
Update iked to now wait on the event loop to exit. Since the main process thread was previously responsible for accepting ipc connections, this code path has been moved to a dedicated thread. This was necessary as the ipc server thread was recently modified to enter an efficient block state while waiting for clients to connect.

git-svn-id: svn://svn.shrew.net/ike@430 1bb156eb-19af-db11-b919-0014c23ffdf2