Re-write portions of the configuration exchange code in iked. The complex send and recv functions have been spit into several smaller functions. These are responsible for client and server operations for xauth, modecfg push or modecfg pull. The client side functions have been tested thoroughly but the server side functions need further review and bug fixing. Additionally, the config state flags have been moved from the tunnel handle to the config handle where they belong.
The major goal of this re-write was to prevent multiple config handles from being used for xauth and modecfg negotiations. Previously, new config handles were created for each unique packet message id. Re-transmitting lost packets was impossible under certain circumstances as the packet queue was destroyed along with the config handle. To avoid this situation, a config handle is now associated with a phase1 handle. The config handle is used to process all config packets for the given phase1 ISAKMP cookie pair. This makes the packet re-transmit queue persistent regardless of the message id. The config handle is destroyed along with the phase1 handle.

git-svn-id: svn://svn.shrew.net/ike@529 1bb156eb-19af-db11-b919-0014c23ffdf2