Rewrite portions of the IPC server code to address some security flaws. Up until now, this code relied on a well behaved client to ensure that certain operations worked correctly. These changes are intended to prevent out of order operations or faulty configurations from triggering invalid pointer references. While addressing the major concerns, more work will be required to validate all configuration options to ensure proper operation in all cases.
Modify iked objects and keyfile code to use BDATA for storage of public and private key information. This removes the openssl key structure pointer from the object and associates functions. Further testing of this change is still required. Also remove the keyfile helper functions from the global iked object. The are only used locally.

Modify the configuration exchange client push and pull functions to copy negotiated option flags to the supported option flags. This allows iked to more accurately track which options were negotiated and which options were statically enabled. The client connect applications need to be updated to reflect this change.

git-svn-id: svn://svn.shrew.net/ike@542 1bb156eb-19af-db11-b919-0014c23ffdf2