Add support for up to four DNS server and two WINS server addresses to the unix Access Manager application. Support multiple name server addresses has existed in iked for quite some time so no changes are required.
Add a new virtual adapter option to the unix Access Manager and VPN Connect applications. This allows a randomized virtual address to be selected from a specified subnet. Using this option has some serious drawbacks. Without the ability to send ARP packets over an IPsec connection, it is impossible to detect and resolve address selection conflicts. However, when a large address pool is used, the odds of multiple clients selecting an identical virtual adapter address is considerably lower than the possibility of multiple clients having identical public address when behind a SOHO router performing NAT. Most of these routers tend to use the same private network definitions by default and are never changed.

Update the unix VPN Connect application to only set the xconf request flag when an option is to be negotiated. Setting the option flag directly denotes that an option is statically configured.


git-svn-id: svn://svn.shrew.net/ike@551 1bb156eb-19af-db11-b919-0014c23ffdf2