Add support to unix gui components for importing cisco PCF files. For PSK based modes, this is a one step process. For RSA based modes, the certificate info must be assigned manually in a second step. Cisco PCF files don't contain certificate data so this cannot be automated.
Modify unix gui to support a new option that allows any remote id and value to be accepted. Although this option should generally be avoided, it was added to help the Cisco PCF import process as they rarely contain remote ID information.

Modify unix gui to support non-address identity types in main mode. Although this option has been requested several times, it technically violates RFCs so I have been reluctant to add it. Now that it is an option, we warn the user when appropriate.


git-svn-id: svn://svn.shrew.net/ike@625 1bb156eb-19af-db11-b919-0014c23ffdf2